server/requirements.txt
openenv-core[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
rdkit>=2023.9.5
python-multipart>=0.0.6
